Added TableStatus
Added a TableStatus property on the Table class. Added a TableStatus
Enum of the same name. In the future, each of these classes should have
a "Status" property and the Enum should represent what kind of Status it
is. e.g. Table.cs should have a Status of TableStatus Enum. Will commit
this later.