Added betting!
Added money properties to player class and win/loss message to Display
class. Also added payout methods to player class.
Few outstanding issues:
-Messages still say that dealer lost or didn't lose money. Should remove
those
-Logic on when to clear Hands should be vetted out. Before, each round
was a new game with new players everytime, but now, we need each Table
instance to have the same players from before, so that bet amounts carry
through. If this is the case, then the hands should be cleared in the
Table Class.
-How will Splits be done? Should hands just "always" be a
List<List<Card>> and we almost always refer to the first List<Card>? or
is there a nicer way of doing this?