Some light refactor to Input/output and Table.PlayerPlay()
Added and implemented an Input class and interface. Added a few more
values to the App.Config. Added player options class. Refactored to the
PlayerPlay class to be a little bit neater. It can probably still use
some work
Issues: Tests do not run anymore because of the app.config. I need to
look into why this is happening