Wooo! Automation
Introduced a new "infinite mode" that allows the user to simply run the
application and the players play and bet for themselves. Obviously the
logic is REALLY simple and possibly flawed in some ways. I should
probably introduce some sort of logging here so I can study the output
later and look for bugs. Also, the normal mode now has 1 player and 3
"bots." so, that's news.